heartbeat/watchdog hot fix

+++ software_architecture_and_design/Server/AppServerMain.cpp
@@ -9,13 +9,21 @@ void fix() {
     if (std::filesystem::exists("resources")) {
         // Remove all files and subdirectories within the folder
         for (const auto& entry : std::filesystem::directory_iterator("resources")) {
-            std::filesystem::remove_all(entry.path());
+            // Check if the filename is "STATE"
+            if (entry.path().filename() != "STATE") {
+                if (std::filesystem::is_directory(entry.path())) {
+                    std::filesystem::remove_all(entry.path());
+                } else {
+                    std::filesystem::remove(entry.path());
+                }
+            }
         }
     } else {
         // Create the directory
         std::filesystem::create_directory("resources");
     }
 }
+
 
 
 int main (int argc, char *argv[])
